Add TS. Add SCSS. Move image to Public. Make Header and Footer part of the Layout. Add branding style for fonts and colors
