+++ package.json
@@ -8,11 +8,17 @@
   },
   "dependencies": {
     "@tailwindcss/forms": "^0.5.3",
+    "@types/jest": "^29.5.1",
+    "@types/node": "^18.16.0",
+    "@types/react": "^18.0.38",
+    "@types/react-dom": "^18.0.11",
     "aos": "^3.0.0-beta.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
-    "react-router-dom": "^6.8.1"
+    "react-router-dom": "^6.8.1",
+    "sass": "^1.62.0",
+    "typescript": "^5.0.4"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^3.1.0",
