--- src/css/additional-styles/hero-slider.css
@@ -1,71 +1,0 @@
-.hero {
-    position: relative;
-    height: 70vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-.hero-slideshow {
-position: absolute;
-top: 0;
-left: 0;
-width: 100%;
-height: 100%;
-overflow: hidden;
-}
-
-.hero-slideshow img {
-position: absolute;
-top: 0;
-left: 0;
-width: 100%;
-height: 100%;
-object-fit: cover;
-transition: opacity 0.5s ease-in-out;
-}
-
-.visible {
-opacity: 1;
-}
-
-.hidden {
-opacity: 0;
-}
-
-.hero-content {
-z-index: 1;
-text-align: center;
-margin: 0 auto;
-max-width: 80%;
-padding: 2rem;
-background-color: #fff;
-box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.hero-content h1 {
-font-size: 3rem;
-margin-bottom: 1rem;
-}
-
-.hero-content p {
-font-size: 1.5rem;
-margin-bottom: 2rem;
-}
-
-.hero-content button {
-padding: 1rem 2rem;
-border-radius: 50px;
-font-size: 1.2rem;
-background-color: #000;
-color: #fff;
-border: none;
-cursor: pointer;
-transition: background-color 0.3s ease-in-out;
-}
-
-.hero-content button:hover {
-background-color: #fff;
-color: #fff;
-}
