+++ src/css/pallete.scss
@@ -1,0 +1,67 @@
+@import url('https://fonts.googleapis.com/css2?family=Black+Han+Sans&family=Poppins:ital,wght@0,400;0,700;1,400&display=swap');
+
+//--- Branding Colors  ---
+$blue: #45acd9;
+$white: #ffffffff;
+$yellow: #ffb400;
+$grey: #474747;
+
+
+//--- Branding Fonts ---
+.headline-1 {
+    font-family: 'Black Han Sans', sans-serif;
+    font-size: 4rem;
+    color: $grey;
+}
+
+.headline-2 {
+    font-family: 'Poppins', sans-serif;
+    font-weight: bold;
+    font-size: 2rem;
+    color: $blue;
+}
+
+.p-font {
+    font-family: 'Poppins', sans-serif;
+    font-weight: 400;
+    font-size: 1rem;
+    color: $grey;
+}
+
+.caption-font {
+    font-family: 'Poppins', sans-serif;
+    font-size: 0.75rem;
+    font-style: italic;
+    color: $grey;
+}
+
+.headline-1-border {
+    position: relative;
+    width: fit-content;
+    padding-left: 1.5rem;
+    padding-right: 1.5rem;
+    border-bottom: 1rem solid  $blue;
+    border-radius: 1rem;
+
+
+   &:before {
+        content : "";
+        position: absolute;
+        right    : 0;
+        z-index: 100;
+        bottom  : 0;
+        width  : 1rem;
+        height   : 25%;  /* or 100px */
+        background: $blue;
+   }
+   &:after {
+        content : "";
+        position: absolute;
+        left    : 0;
+        z-index: 100;
+        bottom  : 0;
+        width  : 1rem;
+        height   : 25%;  /* or 100px */
+        background: $blue;
+   }
+}
