+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+export const decodeEmail = (encodedEmail: string) => {
+    const decoder = new TextDecoder("utf-8");
+    const decodedEmail = decoder.decode(Uint8Array.from(atob(encodedEmail), c => c.charCodeAt(0)));
+    return decodedEmail;
+  }
